+++ lib/libft/src/extension/string/ft_split_len.c
@@ -11,10 +11,11 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stddef.h>
 
-int		ft_split_len(char*const * split)
+size_t		ft_split_len(char*const * split)
 {
-	int len;
+	size_t len;
 
 	len = 0;
 	while (*split)
